Move directMessages react-loadable setup
